Split test def file
